Documentation and small fixes
Documentation and small fixes

+++ include/contacts_db.py
@@ -1,9 +1,26 @@
+"""Constact Application Database API Using SQLite3 Module
+"""
 import os
 import sqlite3 as sl
 from pprint import pprint
 
 class ContactsDB(object):
+
+    """Constact Application Database API
+    
+    Attributes:
+        conn (SQLite Connection Object): Database Connection Object
+        db_path (string): Path to the Database file.
+    """
+    
     def __init__(self, db_path):
+        """Initializes Database.
+        
+        This is either by creating a new SQLite Database file or using an existing one.
+        
+        Args:
+            db_path (string): Path to Local SQLite Database.
+        """
         if db_path != ':memory:':
             if not os.path.isdir(os.path.dirname(db_path)):
                 os.makedirs(os.path.dirname(db_path))
@@ -12,20 +29,40 @@ class ContactsDB(object):
         self.conn = sl.connect(db_path)
         self.conn.row_factory = sl.Row
 
-    def create_table(self, table_name, columns):
-        columns_str = """\nid INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT"""
-        for key, val in columns.items():
+    def create_table(self, table_name, fields):
+        """This will create table <table_name> of specified fields. 
+        
+        An Int type ID field will always be created by default.
+        
+        Args:
+            table_name (string): Table Name.
+            fields (dict): Dictionary where key is the field name and value is field type.
+                            Supported field-type values are (Text, Int, Real, Null).
+        
+        Returns:
+            bool: True if Successful. False if not.
+        
+        Raises:
+            Exception: Invalid Data Type for Field
+        """
+        fields_str = """\nid INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT"""
+        for key, val in fields.items():
             upper_val = val.upper()
-            if upper_val not in ["TEXT", "INT", "INTEGER", "REAL", "NULL"]:
-                self.conn.close()
+            if upper_val not in ["TEXT", "INT", "REAL", "NULL"]:
+                self.close_conn()
                 raise Exception(
-                    "Invalid Data type for Column:",
+                    "Invalid Data type for Field: " + key + ":" + upper_val
                 )
 
-            columns_str += ",\n" + key + " " + upper_val
+            fields_str += ",\n" + key + " " + upper_val
 
-        sql_command = """CREATE TABLE IF NOT EXISTS {0} ({1});""".format(
-            table_name, columns_str
+        tables = self.list_tables()
+        if table_name in tables:
+            self.close_conn()
+            raise Exception("Table:", table_name, "already exists.")
+
+        sql_command = """CREATE TABLE {0} ({1});""".format(
+            table_name, fields_str
         )
         try:
             self.conn.execute(sql_command)
@@ -33,20 +70,44 @@ class ContactsDB(object):
             return True
         except:
             print('Database Operation Failed:', sql_command)
-            self.close()
+            self.close_conn()
             return False
 
     def delete_table(self, table_name):
+        """This will remove table <table_name> if exists.
+        
+        Args:
+            table_name (string): Table Name.
+        
+        Returns:
+            bool: True if Successful. False if not.
+        
+        Raises:
+            Exception: Description
+        """
+        tables = self.list_tables()
+        if table_name not in tables:
+            self.close_conn()
+            raise Exception("Table:" + table_name + " not found.")
+
         sql_command = 'DROP TABLE ' + table_name
         try:
             self.conn.execute(sql_command)
             self.conn.commit()
             return True
         except:
-            self.close()
+            self.close_conn()
             return False
 
     def does_table_exist(self, table_name):
+        """Check if table exists.
+        
+        Args:
+            table_name (string): Table Name.
+        
+        Returns:
+            TYPE: True if it does. False if it doesn't
+        """
         sql_command = (
             "SELECT name FROM sqlite_master WHERE type='table' AND name='{}';".format(
                 table_name
@@ -65,6 +126,11 @@ class ContactsDB(object):
             return None
 
     def list_tables(self):
+        """Display all tables in the Database.
+        
+        Returns:
+            list: list of table names.
+        """
         sql_command = (
             "SELECT name FROM sqlite_master WHERE type='table';")
         try:
@@ -77,6 +143,18 @@ class ContactsDB(object):
             return None
 
     def add(self, table_name, data):
+        """Add Data to Database.
+        
+        Args:
+            table_name (string): Table Name.
+            data (TYPE): Dictionary containing {field:value}
+        
+        Returns:
+            bool: True if Successful. False if not.
+        
+        Raises:
+            Exception: Description
+        """
         if not self.does_table_exist(table_name):
             self.conn.close()
             raise Exception("Table " + table_name + " does not exist.")
@@ -91,13 +169,20 @@ class ContactsDB(object):
 
         key_data = ", ".join(data.keys())
         val_data = ""
+
+        all_fields = self.get_table_fields(table_name)
+
         for key in data.keys():
+            if key not in all_fields:
+                continue
+
+            #  TO DO: Add type-checking for wrong data-type
             if isinstance(data[key], str):
                 val_data += "'" + data[key] + "', "
+
             elif isinstance(data[key], int):
                 val_data += str(data[key]) + ", "
-            else:
-                val_data += "null, "
+
 
         val_data = val_data.rstrip(", ")
         sql_command = "INSERT INTO {0} ({1}) values({2})".format(
@@ -112,6 +197,20 @@ class ContactsDB(object):
             return False
 
     def find(self, table_name, filters, fields, operator="AND"):
+        """Summary
+        
+        Args:
+            table_name (string): Table Name.
+            filters (TYPE): Description
+            fields (TYPE): Description
+            operator (str, optional): Description
+        
+        Returns:
+            TYPE: Description
+        
+        Raises:
+            Exception: Description
+        """
         if not self.does_table_exist(table_name):
             self.conn.close()
             raise Exception("Table " + table_name + " does not exist.")
@@ -120,9 +219,9 @@ class ContactsDB(object):
             self.conn.close()
             raise Exception("Invalid filter data type!")
 
-        columns = self.get_table_columns(table_name)
+        all_fields = self.get_table_fields(table_name)
         if not fields:
-            fields = columns
+            fields = all_fields
 
         field_str = ", ".join(fields)
         fltr_list = []
@@ -133,13 +232,16 @@ class ContactsDB(object):
                 self.conn.close()
                 raise Exception("Invalid filter in filters!")
 
-            if fltr[0] not in columns:
-                self.conn.close()
-                raise Exception("Invalid filter field:" + fltr[0])
+            if fltr[0] not in all_fields:
+                continue
 
+            #  TO DO: Add type-checking for wrong data-type
             if isinstance(fltr[2], str):
                 if fltr[1] == "is":
                     fltr_string = fltr[0] + "='" + fltr[2] + "'"
+
+                elif fltr[1] == "is_not":
+                    fltr_string = "NOT (" + fltr[0] + "='" + fltr[2] + "')"
 
                 elif fltr[1] == "contains":
                     fltr_string = fltr[0] + " LIKE '%" + fltr[2] + "%'"
@@ -151,6 +253,7 @@ class ContactsDB(object):
                     self.conn.close()
                     raise Exception("Invalid Condition for filter:" + str(fltr))
 
+            #  TO DO: Add type-checking for wrong data-type
             if isinstance(fltr[2], int):
                 if fltr[1] == "less_than":
                     fltr_string = fltr[0] + "<" + str(fltr[2])
@@ -162,7 +265,7 @@ class ContactsDB(object):
                     fltr_string = fltr[0] + ">" + str(fltr[2])
 
                 elif fltr[1] == "greater_than_equal":
-                    fltr_string = fltr[0] + "<" + str(fltr[2])
+                    fltr_string = fltr[0] + ">=" + str(fltr[2])
 
                 elif fltr[1] == "equals":
                     fltr_string = fltr[0] + "=" + str(fltr[2])
@@ -202,18 +305,31 @@ class ContactsDB(object):
         return return_data
 
     def update(self, table_name, _id, data):
+        """Update existing data in specified table. Ignores wrong fields.
+        
+        Args:
+            table_name (string): Table Name.
+            _id (int): id of the record.
+            data (dict): New Data record.
+        
+        Returns:
+            bool: True if Successful. False if not.
+        
+        Raises:
+            Exception: Description
+        """
         update_data_str = ""
-        columns = self.get_table_columns(table_name)
+        all_fields = self.get_table_fields(table_name)
 
         for key, val in data.items():
-            if key not in columns:
-                self.conn.close()
-                raise Exception("Invalid Data Field")
+            if key not in all_fields:
+                continue
 
             if key == "id":
                 self.conn.close()
                 raise Exception("ID field is Immutable.")
 
+            #  TO DO: Add type-checking for wrong data-type
             if isinstance(val, str):
                 update_data_str += key + "='" + val + "', "
 
@@ -230,29 +346,40 @@ class ContactsDB(object):
             self.conn.commit()
             return True
         except:
-            self.close()
+            self.close_conn()
             return False
 
-    def delete(self, table_name, ids):
-        if not ids:
-            raise Exception("No ID specified.")
-        if len(ids) == 1:
-            sql_command = (
-                "DELETE FROM " + table_name + " WHERE id=" + str(ids[0]) + ";"
-            )
-        else:
-            sql_command = (
-                "DELETE FROM " + table_name + " WHERE id IN " + str(tuple(ids)) + ";"
-            )
+    def delete(self, table_name, _id):
+        """Delete data record based on specified id.
+        
+        Args:
+            table_name (string): Table Name.
+            _id (int): id of the record.
+        
+        Returns:
+            bool: True if Successful. False if not.
+        """
+        sql_command = (
+            "DELETE FROM " + table_name + " WHERE id=" + str(_id) + ";"
+        )
+        
         try:
             self.conn.execute(sql_command)
             self.conn.commit()
             return True
         except:
-            self.close()
+            self.close_conn()
             return False
 
-    def get_table_columns(self, table_name):
+    def get_table_fields(self, table_name):
+        """Get all fields of specified table name.
+        
+        Args:
+            table_name (string): Table Name.
+        
+        Returns:
+            list: List of field names.
+        """
         sql_command = "SELECT * FROM " + table_name
         try:
             cursor = self.conn.execute(sql_command)
@@ -269,6 +396,14 @@ class ContactsDB(object):
             return names
 
     def clear_all_data(self, table_name):
+        """Clears all data in the specified table.
+        
+        Args:
+            table_name (string): Table Name
+        
+        Returns:
+            bool: True if Successful. False if not.
+        """
         sql_command = (
             "DELETE FROM " + table_name + ";"
         )
@@ -277,10 +412,12 @@ class ContactsDB(object):
             self.conn.commit()
             return True
         except:
-            self.close()
+            self.close_conn()
             return False
 
-    def close(self):
+    def close_conn(self):
+        """Closes Connection.
+        """
         self.conn.close()
